requirements.txt
numpy==1.26.4
sounddevice==0.4.7
soundfile==0.12.1
webrtcvad==2.0.10
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.10.8
rich==13.7.1
pydub==0.25.1
# TTS/ASR providers
azure-cognitiveservices-speech==1.38.0
elevenlabs==1.10.0
# Translators
deepl==1.18.0
deep-translator==1.11.4
# Optional simple UI
streamlit==1.37.1
# Whisper for free ASR
openai-whisper==20240930
